gear

+++ nerago-gearing/main/main.go
@@ -9,6 +9,7 @@ import (
 	"paladin_gearing_go/items"
 	"paladin_gearing_go/model"
 	"paladin_gearing_go/setup"
+	"paladin_gearing_go/simulate"
 	"paladin_gearing_go/util"
 	"runtime/pprof"
 	"time"
@@ -42,9 +43,9 @@ func main() {
 		defer func() {
 			pprof.StopCPUProfile()
 			f.Close()
-			//if simulate.WowSimRanDuringCurrentProcess {
-			os.Rename(f.Name(), files.ProfileDir+"main.pgo")
-			//}
+			if simulate.WowSimRanDuringCurrentProcess {
+				os.Rename(f.Name(), files.ProfileDir+"main.pgo")
+			}
 		}()
 	}
 
@@ -95,9 +96,9 @@ func core(printer *util.PrintRecorder) {
 
 	//statWeightsGrid_updateAll(printer)
 
-	PaladinMultiRun(printer)
+	//PaladinMultiRun(printer)
 
-	//findUpgrades_Paladin(printer)
+	findUpgrades_Paladin(printer)
 }
 
 func setupPallyMitigationSet() (items.FullOptionsMap, model.Model) {
